cleaned code.
